+++ test/i18next_test.dart
@@ -73,9 +73,22 @@ void main() {
     expect(i18next.t('$namespace:myKey'), 'This is my key');
   });
 
-  test('given a non-existing or non matching key', () {
-    expect(i18next.t('someKey'), 'someKey');
-    expect(i18next.t('some.key'), 'some.key');
+  group('given a non-existing or non matching key', () {
+    test('with the default missingKeyHandler', () {
+      expect(i18next.t('someKey'), 'someKey');
+      expect(i18next.t('some.key'), 'some.key');
+    });
+
+    test('with a custom missingKeyHandler', () {
+      const fallbackText = 'Fallback Text';
+      final options = I18NextOptions(
+        missingKeyHandler: expectAsync4(
+          (locale, key, variables, options) => fallbackText,
+          count: 1,
+        ),
+      );
+      expect(i18next.t('someKey', options: options), fallbackText);
+    });
   });
 
   test('given overriding locale', () {
@@ -514,6 +527,20 @@ void main() {
         'fst, snd, and then trd!',
       );
     });
+
+    test(
+      'given a failing interpolation with custom translationFailedHandler',
+      () {
+        const fallbackText = 'Fallback Text';
+        final options = I18NextOptions(
+          translationFailedHandler: expectAsync6(
+            (locale, namespace, key, variables, options, error) => fallbackText,
+            count: 1,
+          ),
+        );
+        expect(i18next.t('$namespace:key', options: options), fallbackText);
+      },
+    );
   });
 
   group('nesting', () {
@@ -621,6 +648,21 @@ void main() {
         mockKey('key', r'My key is $t(key, {"context": "ctx"})!');
 
         expect(i18next.t('$namespace:key'), '$namespace:key');
+      },
+    );
+
+    test(
+      'given a failing nesting with custom translationFailedHandler',
+      () {
+        mockKey('key', r'This is my $t(anotherKey)');
+        const fallbackText = 'Fallback Text';
+        final options = I18NextOptions(
+          translationFailedHandler: expectAsync6(
+            (locale, namespace, key, variables, options, error) => fallbackText,
+            count: 1,
+          ),
+        );
+        expect(i18next.t('$namespace:key', options: options), fallbackText);
       },
     );
   });